Show definitions only in fuzzy search
There is no satisfying way of showing the description
in a way that is segregated from the definition.

Should we define a flag that enables the fuzzy filtering
of the whole snippet (description + definition)?
This would enable filtering based on descriptions.

For now, if users wants to search in descriptions,
they can search directly in the file.