Add empty description by default
If no description is provided, an empty one is added.

This simplifies snippet filtering, as the first
line is always the description.

This greatly speeds up fuzzy filtering eliminating
the visual lag that was visible before. Now there
is no need to spawn a sub-shell for every snippet:
fzf can filter results directly.