Use awk for portability
"--null-data" flag is not supported on default implementation of
sed on Mac OSX. I wouldn't want to add custom logic to the
script to handle this scenario.

Using awk, we get portability and the ability to split the
snippet store into paragraph chunks, that can be separated
by multiple newlines.